+++ app/Http/Controllers/Danhmuc/dmhinhthuclamviecController.php
@@ -5,9 +5,19 @@ namespace App\Http\Controllers\Danhmuc;
 use App\Models\Danhmuc\dmhinhthuclamviec;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Session;
 
 class dmhinhthuclamviecController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (!Session::has('admin')) {
+                return redirect('/home');
+            };
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,20 +25,15 @@ class dmhinhthuclamviecController extends Controller
      */
     public function index()
     {
+        if (!chkPhanQuyen('hinhthuclamviec', 'danhsach')) {
+            return view('errors.noperm')->with('machucnang', 'hinhthuclamviec');
+        }
         $model=dmhinhthuclamviec::all();
         return view('danhmuc.hinhthuclamviec.index')
                 ->with('model',$model);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
+
 
     /**
      * Store a newly created resource in storage.
@@ -38,33 +43,15 @@ class dmhinhthuclamviecController extends Controller
      */
     public function store(Request $request)
     {
+        if (!chkPhanQuyen('hinhthuclamviec', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'hinhthuclamviec');
+        }
         $inputs=$request->all();
         dmhinhthuclamviec::create($inputs);
         return redirect('/dm_hinh_thuc_cong_viec')
                     ->with('success','Thêm mới thành công');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-        //
-    }
 
     /**
      * Update the specified resource in storage.
@@ -75,6 +62,9 @@ class dmhinhthuclamviecController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if (!chkPhanQuyen('hinhthuclamviec', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'hinhthuclamviec');
+        }
         $inputs=$request->all();
         $model=dmhinhthuclamviec::findOrFail($id);
         $model->update($inputs);
@@ -90,6 +80,9 @@ class dmhinhthuclamviecController extends Controller
      */
     public function destroy($id)
     {
+        if (!chkPhanQuyen('hinhthuclamviec', 'thaydoi')) {
+            return view('errors.noperm')->with('machucnang', 'hinhthuclamviec');
+        }
         $model=dmhinhthuclamviec::findOrFail($id);
         $model->delete();
         return redirect('/dm_hinh_thuc_cong_viec')
